+++ packages/eas-cli/src/submit/__tests__/ArchiveSource-test.ts
@@ -9,8 +9,13 @@ import { BuildQuery } from '../../graphql/queries/BuildQuery';
 import { toAppPlatform } from '../../graphql/types/AppPlatform';
 import { confirmAsync, promptAsync } from '../../prompts';
 import { uploadAsync } from '../../uploads';
-import { Archive, ArchiveSourceType, getArchiveAsync } from '../ArchiveSource';
-import { getLatestBuildForSubmissionAsync } from '../utils/builds';
+import {
+  Archive,
+  ArchiveSourceType,
+  BUILD_LIST_ITEM_COUNT,
+  getArchiveAsync,
+} from '../ArchiveSource';
+import { getRecentBuildsForSubmissionAsync } from '../utils/builds';
 
 jest.mock('fs');
 jest.mock('../../log');
@@ -32,6 +37,7 @@ const MOCK_BUILD_FRAGMENT: Partial<BuildFragment> = {
   },
   appVersion: '1.2.3',
   platform: AppPlatform.Android,
+  updatedAt: Date.now(),
 };
 
 const SOURCE_STUB_INPUT = {
@@ -140,7 +146,7 @@ describe(getArchiveAsync, () => {
 
   it('handles latest build source', async () => {
     const projectId = uuidv4();
-    asMock(getLatestBuildForSubmissionAsync).mockResolvedValueOnce(MOCK_BUILD_FRAGMENT);
+    asMock(getRecentBuildsForSubmissionAsync).mockResolvedValueOnce([MOCK_BUILD_FRAGMENT]);
 
     const archive = await getArchiveAsync({
       ...SOURCE_STUB_INPUT,
@@ -148,7 +154,7 @@ describe(getArchiveAsync, () => {
       sourceType: ArchiveSourceType.latest,
     });
 
-    expect(getLatestBuildForSubmissionAsync).toBeCalledWith(
+    expect(getRecentBuildsForSubmissionAsync).toBeCalledWith(
       toAppPlatform(SOURCE_STUB_INPUT.platform),
       projectId
     );
@@ -156,7 +162,7 @@ describe(getArchiveAsync, () => {
   });
 
   it('prompts again if no builds exists when selected latest', async () => {
-    asMock(getLatestBuildForSubmissionAsync).mockResolvedValueOnce(null);
+    asMock(getRecentBuildsForSubmissionAsync).mockResolvedValueOnce([]);
     asMock(promptAsync)
       .mockResolvedValueOnce({ sourceType: ArchiveSourceType.url })
       .mockResolvedValueOnce({ url: ARCHIVE_URL });
@@ -164,6 +170,59 @@ describe(getArchiveAsync, () => {
     const archive = await getArchiveAsync({
       ...SOURCE_STUB_INPUT,
       sourceType: ArchiveSourceType.latest,
+    });
+
+    assertArchiveResult(archive, ArchiveSourceType.url);
+  });
+
+  it('handles build-list-select source', async () => {
+    const projectId = uuidv4();
+    asMock(getRecentBuildsForSubmissionAsync).mockResolvedValueOnce([MOCK_BUILD_FRAGMENT]);
+    asMock(promptAsync).mockResolvedValueOnce({ selectedBuild: MOCK_BUILD_FRAGMENT });
+
+    const archive = await getArchiveAsync({
+      ...SOURCE_STUB_INPUT,
+      projectId,
+      sourceType: ArchiveSourceType.buildList,
+    });
+
+    expect(getRecentBuildsForSubmissionAsync).toBeCalledWith(
+      toAppPlatform(SOURCE_STUB_INPUT.platform),
+      projectId,
+      { limit: BUILD_LIST_ITEM_COUNT }
+    );
+    assertArchiveResult(archive, ArchiveSourceType.buildList);
+  });
+
+  it('prompts again if all builds have expired', async () => {
+    asMock(getRecentBuildsForSubmissionAsync).mockResolvedValueOnce([
+      {
+        ...MOCK_BUILD_FRAGMENT,
+        updatedAt: new Date(Date.now() - 31 * 24 * 3600 * 1000),
+      },
+    ]);
+    asMock(promptAsync)
+      .mockResolvedValueOnce({ sourceType: ArchiveSourceType.url })
+      .mockResolvedValueOnce({ url: ARCHIVE_URL });
+
+    const archive = await getArchiveAsync({
+      ...SOURCE_STUB_INPUT,
+      sourceType: ArchiveSourceType.buildList,
+    });
+
+    assertArchiveResult(archive, ArchiveSourceType.url);
+  });
+
+  it('falls back to prompt if user selected "None of the above"', async () => {
+    asMock(getRecentBuildsForSubmissionAsync).mockResolvedValueOnce([MOCK_BUILD_FRAGMENT]);
+    asMock(promptAsync)
+      .mockResolvedValueOnce({ selectedBuild: null })
+      .mockResolvedValueOnce({ sourceType: ArchiveSourceType.url })
+      .mockResolvedValueOnce({ url: ARCHIVE_URL });
+
+    const archive = await getArchiveAsync({
+      ...SOURCE_STUB_INPUT,
+      sourceType: ArchiveSourceType.buildList,
     });
 
     assertArchiveResult(archive, ArchiveSourceType.url);